refactor(game): consolidate under app/game (#228) (#246)
* refactor(game): consolidate engine/state/rules/types under app/game (#228)

* test(game): assert undo/redo preserves livesRemaining (#228)

* refactor(game): remove deprecated app/_game and wire CI guard (#228)

---------
