+++ __tests__/game/state.test.ts
@@ -1,6 +1,6 @@
-import { initializeGame, applyAction, createEmptyBoard, getCell } from '../../app/_game/state';
-import { isSolved } from '../../app/_game/rules';
-import type { Digit, GameConfig } from '../../app/_game/types';
+import { initializeGame, applyAction, createEmptyBoard, getCell } from '../../app/game/state';
+import { isSolved } from '../../app/game/rules';
+import type { Digit, GameConfig } from '../../app/game/types';
 
 const easyConfig: GameConfig = { difficulty: 'easy', maxLives: 3 };
 
@@ -24,6 +24,18 @@ describe('Game state engine', () => {
     next = applyAction(next, { type: 'erase', row: 1, col: 1 });
     expect(getCell(next.board, 1, 1).value).toBe(null);
     expect(Object.keys(getCell(next.board, 1, 1).notes).length).toBe(0);
+  });
+
+  it('undo/redo do not change livesRemaining', () => {
+    const state = initializeGame([], easyConfig);
+    let next = applyAction(state, { type: 'place', row: 0, col: 0, value: 1 as Digit });
+    // Force a mistake to decrement lives once
+    next = applyAction(next, { type: 'place', row: 0, col: 1, value: 1 as Digit });
+    const livesAfterMistake = next.livesRemaining;
+    const afterUndo = applyAction(next, { type: 'undo' });
+    expect(afterUndo.livesRemaining).toBe(livesAfterMistake);
+    const afterRedo = applyAction(afterUndo, { type: 'redo' });
+    expect(afterRedo.livesRemaining).toBe(livesAfterMistake);
   });
 
   it('isSolved returns true for a solved board', () => {
