--- app/_game/fixtures.ts
@@ -1,1 +1,0 @@
-export * from '../game/fixtures';
